component: Dropdown moved inside hexagon
