+++ src/components/Coaching/WeeksRecords/Table/AssignmentsCell.tsx
@@ -1,11 +1,11 @@
 import type { Assignment, Week } from 'src/types/CoachingTypes';
 import { useAuthAdapter } from '@application/adapters/authAdapter';
+import { Dropdown } from '@interface/components/FormComponents';
 import { useMemo, useState } from 'react';
 import x_dark from 'src/assets/icons/x_dark.svg';
 import {
   CoachDropdown,
   DeleteRecord,
-  Dropdown,
   FormControls,
   LinkInput,
   TextAreaInput,
