+++ src/components/FormComponents/index.ts
@@ -1,7 +1,7 @@
 import Checkbox from './Checkbox';
 import DateInput from './DateInput';
 import DeleteRecord from './DeleteRecord';
-import Dropdown, { CoachDropdown, GenericDropdown } from './Dropdowns';
+import { CoachDropdown, GenericDropdown } from './Dropdowns';
 import FormControls from './FormControls';
 import verifyRequiredInputs from './functions/inputValidation';
 import SearchableDropdown from './SearchableDropdown';
@@ -12,7 +12,6 @@ export {
   CoachDropdown,
   DateInput,
   DeleteRecord,
-  Dropdown,
   FormControls,
   GenericDropdown,
   LinkInput,
