+++ src/components/StudentDrillDown/components/StudentInfoCard.tsx
@@ -1,11 +1,11 @@
 import type { Coach, Student } from 'src/types/CoachingTypes';
 import CoachStudentDrillDown from '@interface/components/CoachStudentDrillDown/CoachStudentDrillDown';
+import { Dropdown } from '@interface/components/FormComponents';
 import React, { useEffect, useState } from 'react';
 import pencil from 'src/assets/icons/pencil.svg';
 import {
   Checkbox,
   CoachDropdown,
-  Dropdown,
   FormControls,
   TextInput,
 } from 'src/components/FormComponents';
