+++ src/hexagon/interface/components/FormComponents/index.ts
@@ -1,0 +1,3 @@
+import Dropdown from '@interface/components/FormComponents/Dropdown/Dropdown';
+
+export { Dropdown };
